requirements.txt
flask==2.3.3
pymongo==3.12.3
flask-bcrypt==1.0.1
flask-session==0.5.0
python-dotenv==1.0.0
dnspython==2.4.2
supabase==2.15.1
certifi==2025.1.31
itsdangerous==2.1.2
openai==1.6.0
requests==2.31.0
Pillow==10.0.0
numpy==1.24.3
pytz==2023.3